@@ -1,12 +1,11 @@
 import Vue from 'vue'
-import Hello from 'src/components/Hello'
+import Hello from '@/components/Hello'
 
 describe('Hello.vue', () => {
   it('should render correct contents', () => {
-    const vm = new Vue({
-      template: '<div><hello></hello></div>',
-      components: { Hello }
-    }).$mount()
-    expect(vm.$el.querySelector('.hello h1').textContent).to.contain('Hello World!')
+    const Constructor = Vue.extend(Hello)
+    const vm = new Constructor().$mount()
+    expect(vm.$el.querySelector('.hello h1').textContent)
+      .to.equal('Welcome to Your Vue.js App')
   })
 })
